+++ src/app/common/enum/query-operators.enum.ts
@@ -1,0 +1,14 @@
+export enum QueryNarrowingOperators {
+  EQ = 'EQ',
+  GT = 'GT',
+  GTE = 'GTE',
+  LT = 'LT',
+  LTE = 'LTE',
+  NE = 'NE',
+  IN = 'IN',
+  NIN = 'NIN',
+  LIKE = 'LIKE',
+  ILIKE = 'ILIKE',
+  CONTAINS = 'CONTAINS',
+  ISNULL = 'ISNULL',
+}